+++ static/scripts/rewards/render-transaction/insert-table-data.ts
@@ -94,6 +94,10 @@ function renderTokenFields(tokenAddress: string, explorerUrl: string) {
 function renderToFields(receiverAddress: string, explorerUrl: string) {
   const toFull = document.querySelector("#rewardRecipient .full") as Element;
   const toShort = document.querySelector("#rewardRecipient .short") as Element;
+
+  // after a single claim toFull returns null as creates a toaster error
+  if(!toFull || !toShort) return console.error("Could not find toFull or toShort");
+
   toFull.innerHTML = `<div>${receiverAddress}</div>`;
   toShort.innerHTML = `<div>${shortenAddress(receiverAddress)}</div>`;
