+++ static/scripts/rewards/web3/erc721-permit.ts
@@ -4,7 +4,7 @@ import { nftRewardAbi } from "../abis/nft-reward-abi";
 import { app } from "../app-state";
 import { renderTransaction } from "../render-transaction/render-transaction";
 import { Erc721Permit } from "../render-transaction/tx-type";
-import { claimButton, errorToast, showLoader, toaster } from "../toaster";
+import { claimButton, showLoader, toaster } from "../toaster";
 import { connectWallet } from "./connect-wallet";
 export function claimErc721PermitHandler(permit: Erc721Permit) {
   return async function claimButtonHandler() {
@@ -46,10 +46,8 @@ export function claimErc721PermitHandler(permit: Erc721Permit) {
         toaster.create("error", `Error rendering transaction: ${error.message}`);
       });
     } catch (error: unknown) {
-      if (error instanceof MetaMaskError) {
-        console.error(error);
-        errorToast(error, error.message ?? error);
-      }
+      console.error(error);
+      toaster.create("error", `Error claiming NFT: ${typeof error === "string" ? error : error.message ? error.message : "Unknown error"}`);
     }
   };
 }
